1차 최종

+++ routes/commentsRouter.js
@@ -5,13 +5,13 @@ const User = require('../schemas/usersSchema');
 const authMiddleware = require('../middlewares/authMiddleware');
 
 // 코멘트 작성: 유저확인
-router.post('/:postId', async (req, res) => {
+router.post('/:postId', authMiddleware, async (req, res) => {
   const { postId } = req.params;
-  //   const { nickname } = res.locals.user;
+  const { nickname } = res.locals.user;
   const { comment } = req.body;
 
-  //   const existingUser = await User.findOne({ nickname: nickname });
-  //   const userImage = existingUser.userImage;
+  const existingUser = await User.findOne({ nickname: nickname });
+  const userImage = existingUser.userImage;
 
   const now = new Date();
   const date = now.toLocaleDateString('ko-KR');
@@ -25,8 +25,8 @@ router.post('/:postId', async (req, res) => {
   const createdComment = await Comment.create({
     comment,
     commentDate,
-    // nickname,
-    // userImage,
+    nickname,
+    userImage,
     postId,
   });
   const commentCount = await Comment.find({ postId: parseInt(postId) });
@@ -42,48 +42,48 @@ router.post('/:postId', async (req, res) => {
 });
 
 // 코멘트 삭제: 유저확인
-router.delete('/:commentId', async (req, res) => {
+router.delete('/:commentId', authMiddleware, async (req, res) => {
   const { commentId } = req.params;
-  //   const { nickname } = res.locals.user;
+  const { nickname } = res.locals.user;
   const existingComment = await Comment.findOne({
     commentId: commentId,
   });
 
-  //   if (nickname !== existingComment.nickname) {
-  //     res
-  //       .status(400)
-  //       .json({ success: false, message: '내가 쓴 댓글이 아닙니다' });
-  //   } else {
-  await Comment.deleteOne({ commentId: parseInt(commentId) });
-  const commentCount = await Comment.find({
-    postId: existingComment.postId,
-  });
-  await Post.updateOne(
-    { postId: existingComment.postId },
-    { $set: { commentCount: commentCount.length } }
-  );
-  res.status(200).json({ success: true, message: '댓글 삭제 성공' });
-  //   }
+  if (nickname !== existingComment.nickname) {
+    res
+      .status(400)
+      .json({ success: false, message: '내가 쓴 댓글이 아닙니다' });
+  } else {
+    await Comment.deleteOne({ commentId: parseInt(commentId) });
+    const commentCount = await Comment.find({
+      postId: existingComment.postId,
+    });
+    await Post.updateOne(
+      { postId: existingComment.postId },
+      { $set: { commentCount: commentCount.length } }
+    );
+    res.status(200).json({ success: true, message: '댓글 삭제 성공' });
+  }
 });
 
 // 코멘트 수정: 유저확인
-router.put('/:commentId', async (req, res) => {
+router.put('/:commentId', authMiddleware, async (req, res) => {
   const { commentId } = req.params;
-  //   const { nickname } = res.locals.user;
+  const { nickname } = res.locals.user;
   const { comment } = req.body;
-  //   const existingComment = await Comment.findOne({ commentId: commentId });
+  const existingComment = await Comment.findOne({ commentId: commentId });
 
-  //   if (nickname !== existingComment.nickname) {
-  //     res
-  //       .status(400)
-  //       .json({ success: false, message: '내가 쓴 댓글이 아닙니다' });
-  //   } else {
-  await Comment.updateOne(
-    { commentId: parseInt(commentId) },
-    { $set: { comment } }
-  );
-  res.status(200).json({ success: true, message: '댓글 수정 완료' });
-  //   }
+  if (nickname !== existingComment.nickname) {
+    res
+      .status(400)
+      .json({ success: false, message: '내가 쓴 댓글이 아닙니다' });
+  } else {
+    await Comment.updateOne(
+      { commentId: parseInt(commentId) },
+      { $set: { comment } }
+    );
+    res.status(200).json({ success: true, message: '댓글 수정 완료' });
+  }
 });
 
 module.exports = router;
